THREADS ARE LIVE BABBY, finished Status of Reader Account , in every 1.mm date reast tables and save values(on mm before date)

+++ Projecton/src/Server/ADailyDBController.java
@@ -4,7 +4,14 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.time.ZonedDateTime;
+import java.util.Calendar;
+import java.util.Date;
 import java.util.Properties;
+import java.util.concurrent.ScheduledThreadPoolExecutor;
+import java.util.concurrent.TimeUnit;
 
 import javax.mail.Message;
 import javax.mail.MessagingException;
@@ -17,12 +24,23 @@ import javax.mail.internet.MimeMessage;
 
 import Common.Mail;
 import Common.ObjectMessage;
+import Common.Report;
 
 public abstract class ADailyDBController 
 {
 	final static String userName = "OBLManager2019@gmail.com";
 	final static String password = "Aa112233";
+    static ScheduledThreadPoolExecutor executor;
+
 	
+    
+    public static void startThreads(Connection connToSQL)
+    {
+    	executor=new ScheduledThreadPoolExecutor(10);
+    	executor.scheduleAtFixedRate(() -> resetUserStatusHistory(connToSQL), 0, 1, TimeUnit.DAYS);
+    }
+    
+    
 	public static ObjectMessage selection(ObjectMessage msg, Connection connToSQL)
 	{
 
@@ -30,15 +48,13 @@ public abstract class ADailyDBController
 		{
 			return sendMail(msg, connToSQL);
 		}
-		else if (((msg.getMessage()).equals("resetUserStatusHistory")))
-		{
-			return resetUserStatusHistory(msg, connToSQL);
-		}
 		else
 		{
 			return null; 
 		}
 	}
+
+	
 	
 	private static ObjectMessage sendMail(ObjectMessage msg, Connection connToSQL) 
 	{
@@ -99,12 +115,11 @@ public abstract class ADailyDBController
 	}
 	
 	//check the methode 
-	private static void writeUserStatusHistory(ObjectMessage msg, Connection connToSQL)
+	private static void monthlyUpdateUserStatusHistory(Connection connToSQL)
 	{
 		PreparedStatement ps=null;
 		ResultSet query1 = null;
 		PreparedStatement ps2=null;
-		ResultSet query2 = null;
 		try 
 		{
 			ps=connToSQL.prepareStatement("SELECT ID, status FROM readeraccount");
@@ -124,29 +139,80 @@ public abstract class ADailyDBController
 		
 	}
 
-	private static ObjectMessage resetUserStatusHistory(ObjectMessage msg, Connection connToSQL)
+	
+	private static void  resetUserStatusHistory(Connection connToSQL)
 	{
-		PreparedStatement ps,ps2;
+		SimpleDateFormat format = new SimpleDateFormat("dd");
+		SimpleDateFormat format2 = new SimpleDateFormat("MM");
+		Date now =  Calendar.getInstance().getTime();
+   	 	String today=format.format(now);
+   	 	if(today.equals("01"))
+   	 	{
+   	 		PreparedStatement ps,ps2;
+   	 		try 
+   	 		{
+   	 			updateHistoryStatusReader(connToSQL,Integer.parseInt(format2.format(now)));
+   	 			ps= connToSQL.prepareStatement("DROP TABLE IF EXISTS `UserStatusHistory`");
+				ps2= connToSQL.prepareStatement("CREATE TABLE `UserStatusHistory` (\r\n" + 
+						"  `readerAccountID` varchar(9) NOT NULL ,\r\n" + 
+						"  `status` enum('Active','Frozen','Locked')  NOT NULL,\r\n" + 
+						"  PRIMARY KEY  (`readerAccountID`,`status`),\r\n" + 
+						"  FOREIGN KEY (`readerAccountID`) REFERENCES `ReaderAccount` (`ID`)\r\n" + 
+						") ENGINE=InnoDB;");
+				ps.executeUpdate();
+				ps2.executeUpdate();
+				monthlyUpdateUserStatusHistory(connToSQL);
+				
+   	 		} 
+   	 		catch (SQLException e) 
+   	 		{
+   	 			e.printStackTrace();
+   	 		}
+   	 	}	
+	}
+
+	
+	private static void  updateHistoryStatusReader(Connection connToSQL,int month)
+	{
+		PreparedStatement ps;
+		ResultSet rs;
+		Report members=new Report();
+		SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
+		Date date = Date.from(ZonedDateTime.now().minusMonths(1).toInstant());
+		String today=sdf.format(date);
 		try 
 		{
-			ps= connToSQL.prepareStatement("DROP TABLE IF EXISTS `UserStatusHistory`");
-			ps2= connToSQL.prepareStatement("CREATE TABLE `UserStatusHistory` (\r\n" + 
-					"  `readerAccountID` varchar(9) NOT NULL ,\r\n" + 
-					"  `status` enum('Active','Frozen','Locked')  NOT NULL,\r\n" + 
-					"  PRIMARY KEY  (`readerAccountID`,`status`),\r\n" + 
-					"  FOREIGN KEY (`readerAccountID`) REFERENCES `ReaderAccount` (`ID`)\r\n" + 
-					") ENGINE=InnoDB;");
+			ps = connToSQL.prepareStatement("SELECT COUNT(*) FROM UserStatusHistory WHERE status=?"); 
+			ps.setString(1, "Active");
+			rs =ps.executeQuery();
+			rs.next();
+			members.setActiveReaderAccounts(Integer.toString(rs.getInt(1)));
+			ps.setString(1, "Frozen");
+			rs =ps.executeQuery();
+			rs.next();
+			members.setFrozenReaderAccounts(Integer.toString(rs.getInt(1)));
+			ps.setString(1, "Locked");
+			rs =ps.executeQuery();
+			rs.next();
+			members.setLockedReaderAccounts(Integer.toString(rs.getInt(1)));
+			ps=connToSQL.prepareStatement("INSERT INTO `History` (`action`,`date`,`Note`) VALUES ('Active readerAccounts',?,?); "); 
+			ps.setString(1, today);
+			ps.setString(2, members.getActiveReaderAccounts());
 			ps.executeUpdate();
-			ps2.executeUpdate();
+			ps=connToSQL.prepareStatement("INSERT INTO `History` (`action`,`date`,`Note`) VALUES ('Frozen readerAccounts',?,?); "); 
+			ps.setString(1, today);
+			ps.setString(2, members.getFrozenReaderAccounts());
+			ps.executeUpdate();
+			ps=connToSQL.prepareStatement("INSERT INTO `History` (`action`,`date`,`Note`) VALUES ('Locked readerAccounts',?,?); "); 
+			ps.setString(1, today);
+			ps.setString(2, members.getLockedReaderAccounts());
+			ps.executeUpdate();
+			
 		} 
 		catch (SQLException e) 
 		{
 			e.printStackTrace();
 		}
-		
-		
-		
-		return null;
 	}
 }
